updated for Xcode6-beta4

+++ Nimble/Adapters/XCTestHandler.swift
@@ -4,7 +4,7 @@ import XCTest
 class XCTestHandler : AssertionHandler {
     func assert(assertion: Bool, message: String, location: SourceLocation) {
         if !assertion {
-            XCTFail(message, file: location.file, line: location.line)
+            XCTFail(message, file: location.file, line: UInt(location.line))
         }
     }
 }
